** I'm going to take a break from front end and begin developing backend.
As of now:

Restart at midnight function is NOT working. I think I'll be able to solve this problem alongside back end development: a lot of functionality in the backend is dependent on this timer so this is a problem that I must solve.

Pop up work, however, they are not customized. This build is without pop ups. There needs to be a total of 8 different pop ups
	- 3 warning users about clicking back ->
	- 3 congratulating users for completing 4 5 6
	- 1 for completing today's challenge + unlocking a new color + timer
	- 1 for completing today's challenge + timer

Also need to add more color assets - Code is ready and in proper reference to these assets.

 I'm happy with the progress made so far, but there is still quite a bit to be done!